Fixed the preservation area lookup and restoration itself.
git-svn-id: https://svn.code.sf.net/p/tsc-tpm/code@250 af914a7b-af0d-0410-90d3-cea0664cafdb

+++ distribution/installation/restore-action.rb
@@ -25,6 +25,8 @@ module Installation
     end
 
     def make_target(progress, logger)
+      return unless File.exists?(source)
+
       FileUtils.remove_entry(target) if File.exists?(target)
       FileUtils.makedirs File.basename(target)
       FileUtils.move source, target
